Add dev-server to help preview catalog (#42)
Adds a new command that spins up a simple dev server that generates and serves the catalogs assets. The server will automatically update when it detects changes in the given catalog repo.

**Drawbacks**

Originally, I had hoped to use a virtual file-system for processing and serving the assets, however I realized it was going to require a bit more refactoring than I wanted to do. Considering the resulting API is ~150MBs, I would recommend mounting a RAM disk and using the --temp-dir flag.

---

Signed-off-by: James Phillips <jamesdphillips@gmail.com>
Co-authored-by: Justin Kolberg <amd.prophet@gmail.com>
Co-authored-by: Sean Porter <portertech@gmail.com>